tests/requirements.txt
boto3
pytest
pytest-mock
python-dotenv
